pyproject.toml
[project]
name = "sql-evaluator"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-core (==0.3.65)",
    "langchain-groq (>=0.3.2,<0.4.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.25,<0.4.0)",
    "gradio (>=5.35.0,<6.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]

[tool.poetry]
packages = [{include = "sql_evaluator", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
